+++ Pepper/Structures/External/FGO/ITraitNameProvider.cs
@@ -2,6 +2,6 @@ namespace Pepper.Structures.External.FGO
 {
     public interface ITraitNameProvider
     {
-        public string GetTrait(long traitId, bool fallbackToEmpty = false);
+        public string GetTrait(int traitId, bool fallbackToEmpty = false);
     }
 }
